dalle/requirements.txt
altgraph==0.17.4
argcomplete==3.1.2
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
idna==3.6
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
pefile==2023.2.7
pipx==1.2.0
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2023.10
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
python-dotenv==1.0.0
pywin32-ctypes==0.2.2
qt-material==2.14
requests==2.31.0
urllib3==2.1.0
userpath==1.9.1
